refactored data into datasets folder + started adding cifar10
