+++ datasets/utils.py
@@ -1,0 +1,13 @@
+import torch
+import numpy as np
+
+def idx_to_float(idx: np.ndarray, num_bins: int):
+    flt_zero_one = (idx + 0.5) / num_bins
+    return (2.0 * flt_zero_one) - 1.0
+
+def float_to_idx(flt: np.ndarray, num_bins: int):
+    flt_zero_one = (flt / 2.0) + 0.5
+    return torch.clamp(torch.floor(flt_zero_one * num_bins), min=0, max=num_bins - 1).long()
+
+def quantize(flt, num_bins: int):
+    return idx_to_float(float_to_idx(flt, num_bins), num_bins)
